Fixing some bugs with bencode parser

+++ src/bencode.h
@@ -12,6 +12,6 @@
 b_dict* parse_bencode_dict(char* input);
 
 b_dict* __parse_dict (char* input, int* position);
-int __parse_int (char* input, int* position);
+int64_t __parse_int (char* input, int* position);
 char* __parse_string (char* input, int* position);
 b_dict_element* __parse_list (char* input, int* position);
